Remove capture of "default" and minor sexp regex cleanup
Before, the keyword control pattern match would match all words starting with "def" (if surrounded by certain characters).
After, the match behaves the same, except it will not match words starting with "default" (including "defaults" "defaulting" etc.). "Default" and its derivatives are considered the only words staring with "def" that would be commonly used in code which are nearly always not definitions.
Also, the sexp regex has been updated to reflect this. While there, I noticed some bad regex (notably, confusion of \w with [a-zA-Z] (which are not the same!) .